REF: UserTerms 관련 코드 별도의 서비스 계층으로 분리
